+++ apps/api/config/routes.rb
@@ -3,6 +3,7 @@ post '/jtb/quote',         to: 'j_t_b#quote'
 post '/kigo/quote',        to: 'kigo#quote'
 post '/kigo/legacy/quote', to: 'kigo/legacy#quote'
 post '/poplidays/quote',   to: 'poplidays#quote'
+post '/ciirus/quote',      to: 'ciirus#quote'
 post '/waytostay/quote',   to: 'waytostay#quote'
 post '/saw/quote',         to: 's_a_w#quote'