+++ apps/api/middlewares/roomorama_webhook.rb
@@ -265,9 +265,22 @@ module API
         status, headers, body = response
         json_response         = json_decode(body.first)
 
-        # this is not a valid Rack response, but the response here is a valid
-        # JSON since it came from Concierge.
-        return false unless json_response.success?
+        # do not touch the response if the route is not found (i.e., status +404+).
+        # In such cases, the logic was all within Hanami and no valid JSON is
+        # returned
+        if status.to_i == 404
+          return response
+        end
+
+        # in case Concierge did not return valid JSON, something is very wrong.
+        # Notify via Rollbar and return an internal server error.
+        unless json_response.success?
+          message = "Invalid JSON from Concierge. Status: #{status}, Body: #{body.first}"
+          error = RuntimeError.new(message)
+
+          Rollbar.error(error)
+          return [500, {}, ["Internal Server Error"]]
+        end
 
         data  = json_response.value
         event = webhook_payload["event"]