+++ spec/api/middlewares/roomorama_webhook_spec.rb
@@ -51,6 +51,8 @@ RSpec.describe API::Middlewares::RoomoramaWebhook do
     }
   }
 
+  subject { described_class.new(upstream) }
+
   describe "invalid webhooks" do
     context "invalid JSON request body" do
       let(:request_body) { StringIO.new("invalid json") }
@@ -209,7 +211,25 @@ RSpec.describe API::Middlewares::RoomoramaWebhook do
     end
   end
 
-  subject { described_class.new(upstream) }
+  context "404" do
+    let(:upstream) { lambda { |env| [404, {}, ["Not Found"]] } }
+
+    it "does not touch the response in case the route is not defined" do
+      expect(post("/invalid", headers)).to eq [404, { "Content-Length" => "9" }, "Not Found"]
+    end
+  end
+
+  context "invalid JSON response from Concierge" do
+    let(:upstream) { lambda { |env| [201, {}, ["Invalid JSON"]] } }
+
+    it "returns a proper internal server error message and reports the occurrence" do
+      error = nil
+      expect(Rollbar).to receive(:error) { |exception| error = exception }
+
+      expect(post("/", headers)).to eq [500, { "Content-Length" => "21" }, "Internal Server Error"]
+      expect(error.message).to eq "Invalid JSON from Concierge. Status: 201, Body: Invalid JSON"
+    end
+  end
 
   def invalid_webhook
     [422, { "Content-Length" => "15" }, "Invalid webhook"]